HSEARCH-3608 Move parameter method to base interface
