+++ mapper/pojo-base/src/main/java/org/hibernate/search/mapper/pojo/bridge/binding/BindingContext.java
@@ -16,4 +16,10 @@ public interface BindingContext {
 	 */
 	BeanResolver beanResolver();
 
+	/**
+	 * @param name The parameter name
+	 * @return Get a parameter defined for the binder by the given name. {@code null} if it does not exist.
+	 */
+	Object parameter(String name);
+
 }
