+++ mapper/pojo-base/src/main/java/org/hibernate/search/mapper/pojo/bridge/binding/IdentifierBindingContext.java
@@ -47,9 +47,4 @@ public interface IdentifierBindingContext<I> extends BindingContext {
 	@Incubating
 	PojoModelValue<I> bridgedElement();
 
-	/**
-	 * @param name The parameter name
-	 * @return Get a parameter defined for the binder by the given name. {@code null} if it does not exist.
-	 */
-	Object parameter(String name);
 }
