+++ mapper/pojo-base/src/main/java/org/hibernate/search/mapper/pojo/bridge/binding/MarkerBindingContext.java
@@ -15,10 +15,4 @@ public interface MarkerBindingContext extends BindingContext {
 	 */
 	void marker(Object marker);
 
-	/**
-	 * @param name The parameter name
-	 * @return Get a parameter defined for the binder by the given name. {@code null} if it does not exist.
-	 */
-	Object parameter(String name);
-
 }
