+++ mapper/pojo-base/src/main/java/org/hibernate/search/mapper/pojo/bridge/binding/PropertyBindingContext.java
@@ -83,10 +83,4 @@ public interface PropertyBindingContext extends BindingContext {
 	 */
 	IndexSchemaElement indexSchemaElement();
 
-	/**
-	 * @param name The parameter name
-	 * @return Get a parameter defined for the binder by the given name. {@code null} if it does not exist.
-	 */
-	Object parameter(String name);
-
 }
