+++ mapper/pojo-base/src/main/java/org/hibernate/search/mapper/pojo/bridge/binding/RoutingBindingContext.java
@@ -44,10 +44,4 @@ public interface RoutingBindingContext extends BindingContext {
 	 */
 	PojoRoutingIndexingDependencyConfigurationContext dependencies();
 
-	/**
-	 * @param name The parameter name
-	 * @return Get a parameter defined for the binder by the given name. {@code null} if it does not exist.
-	 */
-	Object parameter(String name);
-
 }
