+++ mapper/pojo-base/src/main/java/org/hibernate/search/mapper/pojo/bridge/binding/ValueBindingContext.java
@@ -96,10 +96,4 @@ public interface ValueBindingContext<V> extends BindingContext {
 	 */
 	IndexFieldTypeFactory typeFactory();
 
-	/**
-	 * @param name The parameter name
-	 * @return Get a parameter defined for the binder by the given name. {@code null} if it does not exist.
-	 */
-	Object parameter(String name);
-
 }
